+++ lib/db/db-to-gql.d.ts
@@ -19,6 +19,24 @@ export declare class GqlSingleTableInheritanceFactory<TEnum, TGql extends Node<T
     discriminatorKey: string, concreteClassMap: Map<keyof TEnum, ClassType<TGql>>);
     makeFrom(from: TDb, nodeService: NodeService<TGql>): TGql;
     getClassFor(discriminator: keyof TEnum): ClassType<TGql> | undefined;
+    getClasses(): ClassType<TGql>[];
 }
-export declare function modelToClass<T extends Node<T>, V extends Model<V>>(nodeService: NodeService<T>, to: ClassType<T>, from: V, oidScope?: string): T;
+/**
+ * @deprecated for direct use. Use SequelizeBaseServiceInterface.gqlFromDbModel
+ * instead.
+ *
+ * Transforms from a sequelize model to a gql object
+ * THis does not take into account any polymorthic discriminators
+ * and so should not be used directly.
+ *
+ * Note that if any models are eager loaded, they ARE not converted, so the Relay/gql object
+ * references the sequelize model of that name... higher level functions should deal with that
+ * by checkingthe instanceOf the associated model and converting at that time as required.
+ *
+ * @param nodeService
+ * @param to
+ * @param from
+ * @param oidScope
+ */
+export declare function dbToGql<T extends Node<T>, V extends Model<V>>(nodeService: NodeService<T>, to: ClassType<T>, from: V, oidScope?: string): T;
 export declare function reloadNodeFromModel<T extends Node<T>>(node: T, fromDb?: boolean): Promise<T>;